+++ package.json
@@ -10,6 +10,7 @@
     "db:studio": "prisma studio",
     "db:reset": "npx --yes -p dotenv-cli dotenv -c development -- prisma migrate reset",
     "dev": "next dev",
+    "dev:email": "email dev -p 3001",
     "postinstall": "prisma generate",
     "lint": "next lint",
     "start": "next start"
@@ -53,6 +54,7 @@
     "zod": "^3.22.4"
   },
   "devDependencies": {
+    "@react-email/components": "^0.0.17",
     "@types/eslint": "^8.44.7",
     "@types/node": "^18.17.0",
     "@types/react": "^18.2.37",
@@ -65,6 +67,7 @@
     "postcss": "^8.4.31",
     "prettier": "^3.1.0",
     "prettier-plugin-tailwindcss": "^0.5.7",
+    "react-email": "^2.1.2",
     "tailwindcss": "^3.3.5",
     "typescript": "^5.1.6"
   },
