--- react-email/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "emails",
-  "version": "0.0.19",
-  "private": true,
-  "scripts": {
-    "build": "email build",
-    "dev": "email dev -p 3001",
-    "export": "email export"
-  },
-  "dependencies": {
-    "@react-email/components": "0.0.17",
-    "react-email": "2.1.2",
-    "react": "^18.2.0"
-  },
-  "devDependencies": {
-    "@types/react": "18.2.33",
-    "@types/react-dom": "18.2.14"
-  }
-}
