--- react-email/readme.md
@@ -1,27 +1,0 @@
-# React Email Starter
-
-A live preview right in your browser so you don't need to keep sending real emails during development.
-
-## Getting Started
-
-First, install the dependencies:
-
-```sh
-npm install
-# or
-yarn
-```
-
-Then, run the development server:
-
-```sh
-npm run dev
-# or
-yarn dev
-```
-
-Open [localhost:3000](http://localhost:3000) with your browser to see the result.
-
-## License
-
-MIT License
